Create separate Preferences class (node)
