+++ sdks/node-sdk/src/Conversations.ts
@@ -1,5 +1,4 @@
 import type {
-  Consent,
   ConsentState,
   CreateDmOptions,
   CreateGroupOptions,
@@ -12,13 +11,6 @@ import type { Client } from "@/Client";
 import { DecodedMessage } from "@/DecodedMessage";
 import { Dm } from "@/Dm";
 import { Group } from "@/Group";
-
-export type PreferenceUpdate = {
-  type: string;
-  HmacKeyUpdate?: {
-    key: Uint8Array;
-  };
-};
 
 export class Conversations {
   #client: Client;
@@ -284,44 +276,5 @@ export class Conversations {
 
   hmacKeys() {
     return this.#conversations.getHmacKeys();
-  }
-
-  streamConsent(callback?: StreamCallback<Consent[]>) {
-    const asyncStream = new AsyncStream<Consent[]>();
-
-    const stream = this.#conversations.streamConsent((err, value) => {
-      if (err) {
-        asyncStream.callback(err, undefined);
-        callback?.(err, undefined);
-        return;
-      }
-
-      asyncStream.callback(null, value);
-      callback?.(null, value);
-    });
-
-    asyncStream.onReturn = stream.end.bind(stream);
-
-    return asyncStream;
-  }
-
-  streamPreferences(callback?: StreamCallback<PreferenceUpdate>) {
-    const asyncStream = new AsyncStream<PreferenceUpdate>();
-
-    const stream = this.#conversations.streamPreferences((err, value) => {
-      if (err) {
-        asyncStream.callback(err, undefined);
-        callback?.(err, undefined);
-        return;
-      }
-
-      // TODO: remove this once the node bindings type is updated
-      asyncStream.callback(null, value as unknown as PreferenceUpdate);
-      callback?.(null, value as unknown as PreferenceUpdate);
-    });
-
-    asyncStream.onReturn = stream.end.bind(stream);
-
-    return asyncStream;
   }
 }
