+++ sdks/node-sdk/test/Client.test.ts
@@ -1,8 +1,4 @@
-import {
-  ConsentEntityType,
-  ConsentState,
-  IdentifierKind,
-} from "@xmtp/node-bindings";
+import { IdentifierKind } from "@xmtp/node-bindings";
 import { uint8ArrayToHex } from "uint8array-extras";
 import { v4 } from "uuid";
 import { describe, expect, it } from "vitest";
@@ -69,56 +65,6 @@ describe.concurrent("Client", () => {
     expect(inboxId).toBe(client.inboxId);
   });
 
-  it("should return the correct inbox state", async () => {
-    const user = createUser();
-    const signer = createSigner(user);
-    const client = await createRegisteredClient(signer);
-    const inboxState = await client.inboxState();
-    expect(inboxState.inboxId).toBe(client.inboxId);
-    expect(inboxState.installations.map((install) => install.id)).toEqual([
-      client.installationId,
-    ]);
-    expect(inboxState.identifiers).toEqual([await signer.getIdentifier()]);
-    expect(inboxState.recoveryIdentifier).toStrictEqual(
-      await signer.getIdentifier(),
-    );
-
-    const user2 = createUser();
-    const signer2 = createSigner(user2);
-    const client2 = await createClient(signer2);
-    const inboxState2 = await client2.getLatestInboxState(client.inboxId);
-    expect(inboxState2.inboxId).toBe(client.inboxId);
-    expect(inboxState.installations.length).toBe(1);
-    expect(inboxState.installations[0].id).toBe(client.installationId);
-    expect(inboxState2.identifiers).toEqual([await signer.getIdentifier()]);
-    expect(inboxState2.recoveryIdentifier).toStrictEqual(
-      await signer.getIdentifier(),
-    );
-  });
-
-  it("should get inbox states from inbox IDs", async () => {
-    const user = createUser();
-    const user2 = createUser();
-    const signer = createSigner(user);
-    const signer2 = createSigner(user2);
-    const client = await createRegisteredClient(signer);
-    const client2 = await createRegisteredClient(signer2);
-    const inboxStates = await client.inboxStateFromInboxIds([client.inboxId]);
-    expect(inboxStates.length).toBe(1);
-    expect(inboxStates[0].inboxId).toBe(client.inboxId);
-    expect(inboxStates[0].identifiers).toEqual([await signer.getIdentifier()]);
-
-    const inboxStates2 = await client2.inboxStateFromInboxIds(
-      [client2.inboxId],
-      true,
-    );
-    expect(inboxStates2.length).toBe(1);
-    expect(inboxStates2[0].inboxId).toBe(client2.inboxId);
-    expect(inboxStates2[0].identifiers).toEqual([
-      await signer2.getIdentifier(),
-    ]);
-  });
-
   it("should add a wallet association to the client", async () => {
     const user = createUser();
     const user2 = createUser();
@@ -128,7 +74,7 @@ describe.concurrent("Client", () => {
 
     await client.unsafe_addAccount(signer2);
 
-    const inboxState = await client.inboxState();
+    const inboxState = await client.preferences.inboxState();
     expect(inboxState.identifiers.length).toEqual(2);
     expect(inboxState.identifiers).toContainEqual(await signer.getIdentifier());
     expect(inboxState.identifiers).toContainEqual(
@@ -146,7 +92,7 @@ describe.concurrent("Client", () => {
     await client.unsafe_addAccount(signer2);
     await client.removeAccount(await signer2.getIdentifier());
 
-    const inboxState = await client.inboxState();
+    const inboxState = await client.preferences.inboxState();
     expect(inboxState.identifiers).toEqual([await signer.getIdentifier()]);
   });
 
@@ -161,7 +107,7 @@ describe.concurrent("Client", () => {
       dbPath: `./test-${v4()}.db3`,
     });
 
-    const inboxState = await client3.inboxState(true);
+    const inboxState = await client3.preferences.inboxState(true);
     expect(inboxState.installations.length).toBe(3);
 
     const installationIds = inboxState.installations.map((i) => i.id);
@@ -171,7 +117,7 @@ describe.concurrent("Client", () => {
 
     await client3.revokeAllOtherInstallations();
 
-    const inboxState2 = await client3.inboxState(true);
+    const inboxState2 = await client3.preferences.inboxState(true);
 
     expect(inboxState2.installations.length).toBe(1);
     expect(inboxState2.installations[0].id).toBe(client3.installationId);
@@ -188,7 +134,7 @@ describe.concurrent("Client", () => {
       dbPath: `./test-${v4()}.db3`,
     });
 
-    const inboxState = await client3.inboxState(true);
+    const inboxState = await client3.preferences.inboxState(true);
     expect(inboxState.installations.length).toBe(3);
 
     const installationIds = inboxState.installations.map((i) => i.id);
@@ -198,7 +144,7 @@ describe.concurrent("Client", () => {
 
     await client3.revokeInstallations([client.installationIdBytes]);
 
-    const inboxState2 = await client3.inboxState(true);
+    const inboxState2 = await client3.preferences.inboxState(true);
 
     expect(inboxState2.installations.length).toBe(2);
 
@@ -206,45 +152,6 @@ describe.concurrent("Client", () => {
     expect(installationIds2).toContain(client2.installationId);
     expect(installationIds2).toContain(client3.installationId);
     expect(installationIds2).not.toContain(client.installationId);
-  });
-
-  it("should manage consent states", async () => {
-    const user1 = createUser();
-    const user2 = createUser();
-    const signer1 = createSigner(user1);
-    const signer2 = createSigner(user2);
-    const client1 = await createRegisteredClient(signer1);
-    const client2 = await createRegisteredClient(signer2);
-    const group = await client1.conversations.newGroup([client2.inboxId]);
-
-    await client2.conversations.sync();
-    const group2 = await client2.conversations.getConversationById(group.id);
-
-    expect(group2).not.toBeNull();
-
-    expect(
-      await client2.getConsentState(ConsentEntityType.GroupId, group2!.id),
-    ).toBe(ConsentState.Unknown);
-
-    await client2.setConsentStates([
-      {
-        entityType: ConsentEntityType.GroupId,
-        entity: group2!.id,
-        state: ConsentState.Allowed,
-      },
-    ]);
-
-    expect(
-      await client2.getConsentState(ConsentEntityType.GroupId, group2!.id),
-    ).toBe(ConsentState.Allowed);
-
-    expect(group2!.consentState).toBe(ConsentState.Allowed);
-
-    group2!.updateConsentState(ConsentState.Denied);
-
-    expect(
-      await client2.getConsentState(ConsentEntityType.GroupId, group2!.id),
-    ).toBe(ConsentState.Denied);
   });
 
   it("should verify signatures", async () => {
